Fix losing senses that are moved

+++ backend/FwLite/MiniLcm/SyncHelpers/EntrySync.cs
@@ -21,7 +21,9 @@ public static class EntrySync
         Entry[] afterEntries,
         IMiniLcmApi api)
     {
-        return await DiffCollection.DiffAndGetAdded(beforeEntries, afterEntries, new EntriesDiffApi(api));
+        var allBeforeSenses = beforeEntries.SelectMany(e => e.Senses).ToDictionary(s => s.Id);
+        var allAfterSenses = afterEntries.SelectMany(e => e.Senses).ToDictionary(s => s.Id);
+        return await DiffCollection.DiffAndGetAdded(beforeEntries, afterEntries, new EntriesDiffApi(api, allBeforeSenses, allAfterSenses));
     }
 
     /// <summary>
@@ -44,13 +46,13 @@ public static class EntrySync
         return changes;
     }
 
-    public static async Task<int> SyncWithoutComplexFormsAndComponents(Entry beforeEntry, Entry afterEntry, IMiniLcmApi api)
+    public static async Task<int> SyncWithoutComplexFormsAndComponents(Entry beforeEntry, Entry afterEntry, IMiniLcmApi api, Dictionary<Guid, Sense>? allBeforeSenses = null, Dictionary<Guid, Sense>? allAfterSenses = null)
     {
         try
         {
             var updateObjectInput = EntryDiffToUpdate(beforeEntry, afterEntry);
             if (updateObjectInput is not null) await api.UpdateEntry(afterEntry.Id, updateObjectInput);
-            var changes = await SensesSync(afterEntry.Id, beforeEntry.Senses, afterEntry.Senses, api);
+            var changes = await SensesSync(afterEntry.Id, beforeEntry.Senses, afterEntry.Senses, api, allBeforeSenses, allAfterSenses);
             changes += await Sync(afterEntry.Id, beforeEntry.ComplexFormTypes, afterEntry.ComplexFormTypes, api);
             changes += await SyncPublications(afterEntry.Id, beforeEntry.PublishIn, afterEntry.PublishIn, api);
             return changes + (updateObjectInput is null ? 0 : 1);
@@ -119,9 +121,11 @@ public static class EntrySync
     private static async Task<int> SensesSync(Guid entryId,
         IList<Sense> beforeSenses,
         IList<Sense> afterSenses,
-        IMiniLcmApi api)
+        IMiniLcmApi api,
+        Dictionary<Guid, Sense>? allBeforeSenses,
+        Dictionary<Guid, Sense>? allAfterSenses)
     {
-        return await DiffCollection.DiffOrderable(beforeSenses, afterSenses, new SensesDiffApi(api, entryId));
+        return await DiffCollection.DiffOrderable(beforeSenses, afterSenses, new SensesDiffApi(api, entryId, allBeforeSenses, allAfterSenses));
     }
 
     public static UpdateObjectInput<Entry>? EntryDiffToUpdate(Entry beforeEntry, Entry afterEntry)
@@ -137,11 +141,24 @@ public static class EntrySync
         return new UpdateObjectInput<Entry>(patchDocument);
     }
 
-    private class EntriesDiffApi(IMiniLcmApi api) : ObjectWithIdCollectionDiffApi<Entry>
+    private class EntriesDiffApi(IMiniLcmApi api, Dictionary<Guid, Sense> allBeforeSenses, Dictionary<Guid, Sense> allAfterSenses) : ObjectWithIdCollectionDiffApi<Entry>
     {
         public override async Task<(int, Entry)> AddAndGet(Entry afterEntry)
         {
-            var addedEntry = await api.CreateEntry(afterEntry, CreateEntryOptions.WithoutComplexFormsAndComponents);
+            var hasMovedSense = afterEntry.Senses.Any(s => allBeforeSenses.ContainsKey(s.Id));
+            Entry addedEntry;
+            if (hasMovedSense)
+            {
+                // Api.CreateEntry() is optimized and assumes all senses are new rather than moved.
+                // So, we use the "smarter" SensesSync for the senses, which can handle moved senses.
+                addedEntry = await api.CreateEntry(afterEntry with { Senses = [] }, CreateEntryOptions.WithoutComplexFormsAndComponents);
+                await SensesSync(addedEntry.Id, [], afterEntry.Senses, api, allBeforeSenses, allAfterSenses);
+                addedEntry = addedEntry with { Senses = afterEntry.Senses };
+            }
+            else
+            {
+                addedEntry = await api.CreateEntry(afterEntry, CreateEntryOptions.WithoutComplexFormsAndComponents);
+            }
             return (1, addedEntry);
         }
 
@@ -153,7 +170,7 @@ public static class EntrySync
 
         public override Task<int> Replace(Entry before, Entry after)
         {
-            return SyncWithoutComplexFormsAndComponents(before, after, api);
+            return SyncWithoutComplexFormsAndComponents(before, after, api, allBeforeSenses, allAfterSenses);
         }
     }
 
@@ -281,7 +298,7 @@ public static class EntrySync
         }
     }
 
-    private class SensesDiffApi(IMiniLcmApi api, Guid entryId) : IOrderableCollectionDiffApi<Sense, Guid>
+    private class SensesDiffApi(IMiniLcmApi api, Guid entryId, Dictionary<Guid, Sense>? allBeforeSenses, Dictionary<Guid, Sense>? allAfterSenses) : IOrderableCollectionDiffApi<Sense, Guid>
     {
         public Guid GetId(Sense sense)
         {
@@ -290,6 +307,18 @@ public static class EntrySync
 
         public async Task<int> Add(Sense sense, BetweenPosition<Sense> between)
         {
+            /**
+            If we "move sense to different entry" a FieldWorks Lite feature, we could to add "?? await api.GetSense(sense.Id)"
+            when checking for existing senses. But that'd currently just be wasteful and it would probably
+            be better to make moving senses explicit in FieldWorks Lite (i.e. require the client to call MoveSense)
+            */
+            var existing = allBeforeSenses?.GetValueOrDefault(sense.Id);
+            if (existing is not null && existing.EntryId != entryId)
+            {
+                // This can happen when a sense is moved to another entry
+                await api.MoveSense(entryId, sense.Id, new BetweenPosition(between.Previous?.Id, between.Next?.Id));
+                return 1 + await SenseSync.Sync(entryId, existing, sense, api);
+            }
             await api.CreateSense(entryId, sense, new BetweenPosition(between.Previous?.Id, between.Next?.Id));
             return 1;
         }
@@ -302,6 +331,12 @@ public static class EntrySync
 
         public async Task<int> Remove(Sense sense)
         {
+            if (allAfterSenses?.ContainsKey(sense.Id) ?? false)
+            {
+                // Sense wasn't deleted - it was moved to a different entry
+                // We handle that in Add above
+                return 0;
+            }
             await api.DeleteSense(entryId, sense.Id);
             return 1;
         }
